+++ system/lib/commanded/mapper.ex
@@ -4,36 +4,16 @@ defmodule ExESDB.Commanded.Mapper do
   """
   alias Commanded.EventStore.EventData, as: EventData
   alias Commanded.EventStore.RecordedEvent, as: RecordedEvent
+  alias Commanded.EventStore.SnapshotData, as: SnapshotData
+
   alias ExESDB.EventRecord, as: EventRecord
   alias ExESDB.NewEvent, as: NewEvent
+  alias ExESDB.Schema.SnapshotRecord, as: SnapshotRecord
 
   require UUIDv7
 
   @doc """
     Converts a Commanded EventData struct to an ExESDB NewEvent struct.
-    
-    Example:
-    iex> event_data = %Commanded.EventStore.EventData{
-    ...>   event_type: "OrderPlaced",
-    ...>   data: %{order_id: "123", customer_id: "456"},
-    ...>   metadata: %{correlation_id: "123", causation_id: "456"},
-    ...>   correlation_id: "123",
-    ...>   causation_id: "456"
-    ...> }
-    iex> event_id = UUIDv7.generate()
-    iex> ExESDB.Mapper.to_new_event(event_data)
-    %NewEvent{
-      event_id: ^event_id,
-      event_type: "OrderPlaced",
-      data_content_type: 1,
-      metadata_content_type: 1,
-      data: %{order_id: "123", customer_id: "456"},
-      metadata: %{
-        correlation_id: "123",
-        causation_id: "456"
-        stream_version: 0
-      }
-    }
   """
   @spec to_new_event(EventData.t()) :: NewEvent.t()
   def to_new_event(event_data)
@@ -53,53 +33,6 @@ defmodule ExESDB.Commanded.Mapper do
 
   @doc """
     Converts an ExESDB EventRecord struct to a Commanded RecordedEvent struct.
-
-    Example:
-    iex> event_stream_id = "stream-1"
-    ^event_stream_id
-    iex> event_id = UUIDv7.generate()
-    ^event_id
-    iex> created = DateTime.utc_now()
-    ^created
-    iex> created_epoch = DateTime.to_unix(created, :millisecond)
-    ^created_epoch
-    iex> event_record = %EventRecord{
-    ...>   event_stream_id: "stream-1",
-    ...>   event_number: 1,
-    ...>   event_id: ^event_id,
-    ...>   event_type: "OrderPlaced",
-    ...>   data_content_type: 1,
-    ...>   metadata_content_type: 1,
-    ...>   data: %{order_id: "123", customer_id: "456"},
-    ...>   metadata: %{
-    ...>     correlation_id: "123",
-    ...>     causation_id: "456",
-    ...>     stream_version: 0
-    ...>   },
-    ...>   created: ^created,
-    ...>   created_epoch: ^created_epoch
-    ...> }
-    ^event_record
-    iex> ExESDB.Mapper.to_recorded_event(event_record)
-    %RecordedEvent{
-      event_id: ^event_id,
-      event_number: 1,
-      event_type: "OrderPlaced",
-      data: %{
-        order_id: "123", 
-        customer_id: "456"
-      },
-      metadata: %{
-        correlation_id: "123",
-        causation_id: "456",
-        stream_version: 0
-      },
-      created_at: ^created,
-      stream_id: "stream-1",
-      stream_version: 0,
-      correlation_id: "123",
-      causation_id: "456"
-    }
   """
   @spec to_recorded_event(EventRecord.t()) :: RecordedEvent.t()
   def to_recorded_event(
@@ -123,5 +56,35 @@ defmodule ExESDB.Commanded.Mapper do
         stream_version: stream_version,
         correlation_id: correlation_id,
         causation_id: causation_id
+      }
+
+  @doc """
+    Converts an Commanded SnapshotData struct to an ExESDB SnapshotRecord struct.
+  """
+  @spec to_snapshot_record(SnapshotData.t()) :: SnapshotRecord.t()
+  def to_snapshot_record(snapshot_data)
+      when is_struct(snapshot_data, SnapshotData),
+      do: %SnapshotRecord{
+        source_uuid: snapshot_data.source_uuid,
+        source_version: snapshot_data.source_version,
+        source_type: snapshot_data.source_type,
+        data: snapshot_data.data,
+        metadata: snapshot_data.metadata,
+        created_at: snapshot_data.created_at,
+        created_epoch: DateTime.to_unix(snapshot_data.created_at, :millisecond)
+      }
+
+  @doc """
+    Converts an ExESDB SnapshotRecord struct to a Commanded SnapshotData struct.
+  """
+  def to_snapshot_data(snapshot_record)
+      when is_struct(snapshot_record, SnapshotRecord),
+      do: %SnapshotData{
+        source_uuid: snapshot_record.source_uuid,
+        source_version: snapshot_record.source_version,
+        source_type: snapshot_record.source_type,
+        data: snapshot_record.data,
+        metadata: snapshot_record.metadata,
+        created_at: snapshot_record.created_at
       }
 end